split arbiter ui in to its own module
