+++ arbiter-core/pom.xml
@@ -62,12 +62,6 @@
         </dependency>
 
 
-        <dependency>
-            <groupId>org.deeplearning4j</groupId>
-            <artifactId>deeplearning4j-ui-components</artifactId>
-            <version>${dl4j.version}</version>
-        </dependency>
-
 
 
         <!-- ND4J Shaded Jackson Dependency -->
