+++ arbiter-deeplearning4j-ui/src/main/java/org/deeplearning4j/arbiter/optimize/resources/SummaryStatusResource.java
@@ -15,7 +15,7 @@
  *  *    limitations under the License.
  *
  */
-package org.deeplearning4j.arbiter.optimize.ui.resources;
+package org.deeplearning4j.arbiter.optimize.resources;
 
 import org.deeplearning4j.arbiter.optimize.serde.jackson.JsonMapper;
 import org.deeplearning4j.ui.api.Component;
@@ -28,26 +28,27 @@ import javax.ws.rs.core.Response;
 import java.util.Collections;
 import java.util.concurrent.atomic.AtomicInteger;
 
-@Path("/config")
+/**Summary stats: number of completed tasks etc
+ */
+@Path("/summary")
 @Produces(MediaType.APPLICATION_JSON)
-public class ConfigResource {
-
-    public static final Logger log = LoggerFactory.getLogger(ConfigResource.class);
-    private Component component = null;
+public class SummaryStatusResource {
+    public static Logger log = LoggerFactory.getLogger(SummaryStatusResource.class);
 
     private static final int maxWarnCount = 5;
     private AtomicInteger warnCount = new AtomicInteger(0);
 
-    @GET
-    public Response getConfig(){
-        log.trace("GET for config with current component: {}",component);
+    private Component component = null;
 
+    @GET
+    public Response getStatus(){
+        log.trace("Get with elements: {}",component);
         String str = "";
         try{
             str = JsonMapper.getMapper().writeValueAsString(component);
         } catch (Exception e){
             if(warnCount.getAndIncrement() < maxWarnCount){
-                log.warn("Error getting configuration UI info", e);
+                log.warn("Error getting summary status update", e);
             }
         }
         Response r = Response.ok(str).build();
@@ -56,20 +57,20 @@ public class ConfigResource {
 
     @POST
     @Path("/update")
-    @Consumes(MediaType.TEXT_PLAIN)
     @Produces(MediaType.APPLICATION_JSON)
-    public Response update(String componentStr){
-        log.trace("POST for config with component: {}",component);
+    @Consumes(MediaType.TEXT_PLAIN)
+    public Response update(String component){
+        log.trace("Post with new elements: {}",component);
 
-        if(componentStr == null || componentStr.isEmpty()){
+        if(component == null || component.isEmpty()){
             return Response.ok(Collections.singletonMap("status", "ok")).build();
         }
 
         try{
-            this.component = JsonMapper.getMapper().readValue(componentStr, Component.class);
+            this.component = JsonMapper.getMapper().readValue(component, Component.class);
         } catch (Exception e) {
             if(warnCount.getAndIncrement() < maxWarnCount){
-                log.warn("Error posting config stats update", e);
+                log.warn("Error posting summary status update", e);
             }
         }
