+++ arbiter-deeplearning4j-ui/src/main/java/org/deeplearning4j/arbiter/optimize/resources/ui/ClientProvider.java
@@ -15,7 +15,7 @@
  *  *    limitations under the License.
  *
  */
-package org.deeplearning4j.arbiter.optimize.ui;
+package org.deeplearning4j.arbiter.optimize.resources.ui;
 
 import com.fasterxml.jackson.jaxrs.json.JacksonJsonProvider;
