+++ arbiter-deeplearning4j-ui/src/main/java/org/deeplearning4j/arbiter/optimize/runner/listener/candidate/UICandidateStatusListener.java
