+++ arbiter-deeplearning4j-ui/src/main/java/org/deeplearning4j/arbiter/optimize/runner/listener/candidate/UICandidateStatusListenerImpl.java
@@ -18,9 +18,8 @@
 package org.deeplearning4j.arbiter.optimize.runner.listener.candidate;
 
 import org.deeplearning4j.arbiter.optimize.runner.Status;
-import org.deeplearning4j.arbiter.optimize.ui.ArbiterUIServer;
-import org.deeplearning4j.arbiter.optimize.ui.ClientProvider;
 import org.deeplearning4j.arbiter.optimize.serde.jackson.JsonMapper;
+
 import org.deeplearning4j.ui.api.Component;
 import org.deeplearning4j.ui.api.LengthUnit;
 import org.deeplearning4j.ui.components.component.ComponentDiv;
@@ -49,7 +48,6 @@ public class UICandidateStatusListenerImpl implements UICandidateStatusListener 
 
     public UICandidateStatusListenerImpl(int candidateNumber){
         this.candidateNumber = candidateNumber;
-        target = ClientProvider.getClient().target("http://localhost:" + ArbiterUIServer.getInstance().getPort() + "/modelResults/update/" + candidateNumber);
     }
