+++ arbiter-deeplearning4j/pom.xml
@@ -27,11 +27,7 @@
             <version>${dl4j.version}</version>
         </dependency>
 
-        <dependency>
-            <groupId>org.deeplearning4j</groupId>
-            <artifactId>deeplearning4j-ui-components</artifactId>
-            <version>${dl4j.version}</version>
-        </dependency>
+
 
         <dependency>
             <groupId>org.nd4j</groupId>
