+++ arbiter-deeplearning4j/src/main/java/org/deeplearning4j/arbiter/task/ComputationGraphTaskCreator.java
@@ -18,23 +18,18 @@
 package org.deeplearning4j.arbiter.task;
 
 import lombok.AllArgsConstructor;
-import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.deeplearning4j.arbiter.GraphConfiguration;
-import org.deeplearning4j.arbiter.listener.UIGraphStatusReportingListener;
 import org.deeplearning4j.arbiter.optimize.api.Candidate;
 import org.deeplearning4j.arbiter.optimize.api.OptimizationResult;
 import org.deeplearning4j.arbiter.optimize.api.TaskCreator;
 import org.deeplearning4j.arbiter.optimize.api.data.DataProvider;
 import org.deeplearning4j.arbiter.optimize.api.evaluation.ModelEvaluator;
 import org.deeplearning4j.arbiter.optimize.api.score.ScoreFunction;
-import org.deeplearning4j.arbiter.optimize.runner.Status;
-import org.deeplearning4j.arbiter.optimize.runner.listener.candidate.UICandidateStatusListener;
 import org.deeplearning4j.arbiter.scoring.util.ScoreUtil;
 import org.deeplearning4j.earlystopping.EarlyStoppingConfiguration;
 import org.deeplearning4j.earlystopping.EarlyStoppingResult;
 import org.deeplearning4j.earlystopping.trainer.EarlyStoppingGraphTrainer;
 import org.deeplearning4j.nn.graph.ComputationGraph;
-import org.deeplearning4j.ui.components.text.ComponentText;
 import org.nd4j.linalg.dataset.api.iterator.DataSetIterator;
 
 import java.util.concurrent.Callable;
@@ -46,17 +41,16 @@ import java.util.concurrent.Callable;
  * @author Alex Black
  */
 @AllArgsConstructor
-public class ComputationGraphTaskCreator<A> implements TaskCreator<GraphConfiguration, ComputationGraph, Object, A> {
+public class ComputationGraphTaskCreator<A> implements TaskCreator<GraphConfiguration, ComputationGraph, Object,A> {
 
     private ModelEvaluator<ComputationGraph, Object, A> modelEvaluator;
 
     @Override
     public Callable<OptimizationResult<GraphConfiguration, ComputationGraph, A>> create(
             Candidate<GraphConfiguration> candidate, DataProvider<Object> dataProvider,
-            ScoreFunction<ComputationGraph, Object> scoreFunction,
-            UICandidateStatusListener statusListener) {
+            ScoreFunction<ComputationGraph, Object> scoreFunction) {
 
-        return new GraphLearningTask<>(candidate, dataProvider, scoreFunction, modelEvaluator, statusListener);
+        return new GraphLearningTask<>(candidate, dataProvider, scoreFunction, modelEvaluator);
     }
 
 
@@ -67,18 +61,14 @@ public class ComputationGraphTaskCreator<A> implements TaskCreator<GraphConfigur
         private ScoreFunction<ComputationGraph, Object> scoreFunction;
         private ModelEvaluator<ComputationGraph, Object, A> modelEvaluator;
 
-        private UIGraphStatusReportingListener dl4jListener;
 
         public GraphLearningTask(Candidate<GraphConfiguration> candidate, DataProvider<Object> dataProvider,
                                  ScoreFunction<ComputationGraph, Object> scoreFunction,
-                                 ModelEvaluator<ComputationGraph, Object, A> modelEvaluator,
-                                 UICandidateStatusListener listener) {
+                                 ModelEvaluator<ComputationGraph, Object, A> modelEvaluator) {
             this.candidate = candidate;
             this.dataProvider = dataProvider;
             this.scoreFunction = scoreFunction;
             this.modelEvaluator = modelEvaluator;
-
-            dl4jListener = new UIGraphStatusReportingListener(listener);
         }
 
 
@@ -87,7 +77,6 @@ public class ComputationGraphTaskCreator<A> implements TaskCreator<GraphConfigur
             //Create network
             ComputationGraph net = new ComputationGraph(candidate.getValue().getConfiguration());
             net.init();
-            net.setListeners(dl4jListener);
 
             //Early stopping or fixed number of epochs:
             DataSetIterator dataSetIterator = ScoreUtil.getIterator(dataProvider.trainData(candidate.getDataParameters()));
@@ -96,24 +85,24 @@ public class ComputationGraphTaskCreator<A> implements TaskCreator<GraphConfigur
             EarlyStoppingConfiguration<ComputationGraph> esConfig = candidate.getValue().getEarlyStoppingConfiguration();
             EarlyStoppingResult<ComputationGraph> esResult = null;
             if (esConfig != null) {
-                EarlyStoppingGraphTrainer trainer = new EarlyStoppingGraphTrainer(esConfig, net, dataSetIterator, dl4jListener);
+                EarlyStoppingGraphTrainer trainer = new EarlyStoppingGraphTrainer(esConfig, net, dataSetIterator);
                 try {
                     esResult = trainer.fit();
                     net = esResult.getBestModel();  //Can return null if failed OR if
                 } catch (Exception e) {
-                    dl4jListener.postReport(Status.Failed, null,
+                    /*dl4jListener.postReport(Status.Failed, null,
                             new ComponentText("Unexpected exception during model training\n", null),
-                            new ComponentText(ExceptionUtils.getStackTrace(e), null));
+                            new ComponentText(ExceptionUtils.getStackTrace(e), null));*/
                     throw e;
                 }
 
                 switch (esResult.getTerminationReason()) {
                     case Error:
-                        dl4jListener.postReport(Status.Failed, esResult);
+                      //  dl4jListener.postReport(Status.Failed, esResult);
                         break;
                     case IterationTerminationCondition:
                     case EpochTerminationCondition:
-                        dl4jListener.postReport(Status.Complete, esResult);
+                      //  dl4jListener.postReport(Status.Complete, esResult);
                         break;
                 }
 
@@ -125,7 +114,7 @@ public class ComputationGraphTaskCreator<A> implements TaskCreator<GraphConfigur
                     dataSetIterator.reset();
                 }
                 //Do a final status update
-                dl4jListener.postReport(Status.Complete, null);
+               // dl4jListener.postReport(Status.Complete, null);
             }
 
             A additionalEvaluation = null;
@@ -133,23 +122,25 @@ public class ComputationGraphTaskCreator<A> implements TaskCreator<GraphConfigur
                 try {
                     additionalEvaluation = (modelEvaluator != null ? modelEvaluator.evaluateModel(net, dataProvider) : null);
                 } catch (Exception e) {
-                    dl4jListener.postReport(Status.Failed, esResult,
+                    e.printStackTrace();
+                  /*  dl4jListener.postReport(Status.Failed, esResult,
                             new ComponentText("Failed during additional evaluation stage\n", null),
-                            new ComponentText(ExceptionUtils.getStackTrace(e), null));
+                            new ComponentText(ExceptionUtils.getStackTrace(e), null));*/
                 }
             }
 
             Double score = null;
             if (net == null) {
-                dl4jListener.postReport(Status.Complete, esResult,
-                        new ComponentText("No best model available; cannot calculate model score", null));
+               /* dl4jListener.postReport(Status.Complete, esResult,
+                        new ComponentText("No best model available; cannot calculate model score", null));*/
             } else {
                 try {
                     score = scoreFunction.score(net, dataProvider, candidate.getDataParameters());
                 } catch (Exception e) {
-                    dl4jListener.postReport(Status.Failed, esResult,
+                    e.printStackTrace();
+                   /* dl4jListener.postReport(Status.Failed, esResult,
                             new ComponentText("Failed during score calculation stage\n", null),
-                            new ComponentText(ExceptionUtils.getStackTrace(e), null));
+                            new ComponentText(ExceptionUtils.getStackTrace(e), null));*/
                 }
             }
