+++ arbiter-deeplearning4j/src/main/java/org/deeplearning4j/arbiter/task/MultiLayerNetworkTaskCreator.java
@@ -19,26 +19,18 @@ package org.deeplearning4j.arbiter.task;
 
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
-import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.deeplearning4j.arbiter.DL4JConfiguration;
-import org.deeplearning4j.arbiter.listener.BaseUIStatusReportingListener;
-import org.deeplearning4j.arbiter.listener.UIStatusReportingListener;
 import org.deeplearning4j.arbiter.optimize.api.Candidate;
 import org.deeplearning4j.arbiter.optimize.api.OptimizationResult;
 import org.deeplearning4j.arbiter.optimize.api.TaskCreator;
 import org.deeplearning4j.arbiter.optimize.api.data.DataProvider;
 import org.deeplearning4j.arbiter.optimize.api.evaluation.ModelEvaluator;
 import org.deeplearning4j.arbiter.optimize.api.score.ScoreFunction;
-import org.deeplearning4j.arbiter.optimize.runner.Status;
-import org.deeplearning4j.arbiter.optimize.runner.listener.candidate.UICandidateStatusListener;
-
-import org.deeplearning4j.arbiter.optimize.ui.ArbiterUIServer;
 import org.deeplearning4j.arbiter.scoring.util.ScoreUtil;
 import org.deeplearning4j.earlystopping.EarlyStoppingConfiguration;
 import org.deeplearning4j.earlystopping.EarlyStoppingResult;
 import org.deeplearning4j.earlystopping.trainer.EarlyStoppingTrainer;
 import org.deeplearning4j.nn.multilayer.MultiLayerNetwork;
-import org.deeplearning4j.ui.components.text.ComponentText;
 import org.nd4j.linalg.dataset.api.iterator.DataSetIterator;
 
 import java.util.concurrent.Callable;
@@ -58,10 +50,9 @@ public class MultiLayerNetworkTaskCreator<A> implements TaskCreator<DL4JConfigur
     @Override
     public Callable<OptimizationResult<DL4JConfiguration, MultiLayerNetwork, A>> create(
             Candidate<DL4JConfiguration> candidate, DataProvider<Object> dataProvider,
-            ScoreFunction<MultiLayerNetwork, Object> scoreFunction,
-            UICandidateStatusListener statusListener) {
+            ScoreFunction<MultiLayerNetwork, Object> scoreFunction) {
 
-        return new DL4JLearningTask<>(candidate, dataProvider, scoreFunction, modelEvaluator, statusListener);
+        return new DL4JLearningTask<>(candidate, dataProvider, scoreFunction, modelEvaluator);
 
     }
 
@@ -73,19 +64,15 @@ public class MultiLayerNetworkTaskCreator<A> implements TaskCreator<DL4JConfigur
         private ScoreFunction<MultiLayerNetwork, Object> scoreFunction;
         private ModelEvaluator<MultiLayerNetwork, Object, A> modelEvaluator;
 
-        private BaseUIStatusReportingListener<MultiLayerNetwork> dl4jListener;
 
         public DL4JLearningTask(Candidate<DL4JConfiguration> candidate,
                                 DataProvider<Object> dataProvider,
                                 ScoreFunction<MultiLayerNetwork, Object> scoreFunction,
-                                ModelEvaluator<MultiLayerNetwork, Object, A> modelEvaluator,
-                                UICandidateStatusListener listener) {
+                                ModelEvaluator<MultiLayerNetwork, Object, A> modelEvaluator) {
             this.candidate = candidate;
             this.dataProvider = dataProvider;
             this.scoreFunction = scoreFunction;
             this.modelEvaluator = modelEvaluator;
-
-            dl4jListener = (ArbiterUIServer.isRunning() ? new UIStatusReportingListener(listener) : null);
         }
 
 
@@ -94,7 +81,6 @@ public class MultiLayerNetworkTaskCreator<A> implements TaskCreator<DL4JConfigur
             //Create network
             MultiLayerNetwork net = new MultiLayerNetwork(candidate.getValue().getMultiLayerConfiguration());
             net.init();
-            net.setListeners(dl4jListener);
 
             //Early stopping or fixed number of epochs:
             DataSetIterator dataSetIterator = ScoreUtil.getIterator(dataProvider.trainData(candidate.getDataParameters()));
@@ -103,31 +89,19 @@ public class MultiLayerNetworkTaskCreator<A> implements TaskCreator<DL4JConfigur
             EarlyStoppingConfiguration<MultiLayerNetwork> esConfig = candidate.getValue().getEarlyStoppingConfiguration();
             EarlyStoppingResult<MultiLayerNetwork> esResult = null;
             if (esConfig != null) {
-                EarlyStoppingTrainer trainer = new EarlyStoppingTrainer(esConfig, net, dataSetIterator, dl4jListener);
+                EarlyStoppingTrainer trainer = new EarlyStoppingTrainer(esConfig, net, dataSetIterator);
                 try {
                     esResult = trainer.fit();
                     net = esResult.getBestModel();  //Can return null if failed OR if
                 } catch (Exception e) {
-                    if(dl4jListener != null) {
-                        dl4jListener.postReport(Status.Failed, null,
-                                new ComponentText("Unexpected exception during model training\n", null),
-                                new ComponentText(ExceptionUtils.getStackTrace(e), null));
-                    }
                     throw e;
                 }
 
                 switch (esResult.getTerminationReason()) {
                     case Error:
-                        if(dl4jListener != null) {
-                            dl4jListener.postReport(Status.Failed, esResult);
-                        }
                         break;
                     case IterationTerminationCondition:
-                    case EpochTerminationCondition:
-                        if(dl4jListener != null) {
-                            dl4jListener.postReport(Status.Complete, esResult);
-                        }
-                        break;
+                    case EpochTerminationCondition: break;
                 }
 
             } else {
@@ -137,10 +111,6 @@ public class MultiLayerNetworkTaskCreator<A> implements TaskCreator<DL4JConfigur
                     net.fit(dataSetIterator);
                     dataSetIterator.reset();
                 }
-                //Do a final status update
-                if(dl4jListener != null) {
-                    dl4jListener.postReport(Status.Complete, null);
-                }
             }
 
             A additionalEvaluation = null;
@@ -148,29 +118,15 @@ public class MultiLayerNetworkTaskCreator<A> implements TaskCreator<DL4JConfigur
                 try {
                     additionalEvaluation = (modelEvaluator != null ? modelEvaluator.evaluateModel(net, dataProvider) : null);
                 } catch (Exception e) {
-                    if(dl4jListener != null) {
-                        dl4jListener.postReport(Status.Failed, esResult,
-                                new ComponentText("Failed during additional evaluation stage\n", null),
-                                new ComponentText(ExceptionUtils.getStackTrace(e), null));
-                    }
+                    e.printStackTrace();
                 }
             }
 
             Double score = null;
             if (net == null) {
-                if(dl4jListener != null) {
-                    dl4jListener.postReport(Status.Complete, esResult,
-                            new ComponentText("No best model available; cannot calculate model score", null));
-                }
-            } else {
                 try {
                     score = scoreFunction.score(net, dataProvider, candidate.getDataParameters());
                 } catch (Exception e) {
-                    if(dl4jListener != null) {
-                        dl4jListener.postReport(Status.Failed, esResult,
-                                new ComponentText("Failed during score calculation stage\n", null),
-                                new ComponentText(ExceptionUtils.getStackTrace(e), null));
-                    }
                 }
             }
