+++ pom.xml
@@ -23,6 +23,7 @@
         <module>arbiter-deeplearning4j</module>
         <module>arbiter-core</module>
         <module>arbiter-server</module>
+        <module>arbiter-deeplearning4j-ui</module>
     </modules>
     <packaging>pom</packaging>
     <name>Arbiter</name>
